move : BaseEntity 파일 이동 및 by 계열 속성 Long으로 변경
