+++ src/main/java/com/example/team1_be/domain/Schedule/Schedule.java
@@ -1,7 +1,7 @@
 package com.example.team1_be.domain.Schedule;
 
 import com.example.team1_be.domain.Group.Group;
-import com.example.team1_be.utils.BaseEntity;
+import com.example.team1_be.utils.audit.BaseEntity;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
