+++ src/main/java/com/example/team1_be/domain/Substitute/Substitute.java
@@ -1,7 +1,7 @@
 package com.example.team1_be.domain.Substitute;
 
 import com.example.team1_be.domain.Apply.Apply;
-import com.example.team1_be.utils.BaseEntity;
+import com.example.team1_be.utils.audit.BaseEntity;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
