+++ src/main/java/com/example/team1_be/domain/User/User.java
@@ -1,7 +1,7 @@
 package com.example.team1_be.domain.User;
 
 import com.example.team1_be.domain.Member.Member;
-import com.example.team1_be.utils.BaseEntity;
+import com.example.team1_be.utils.audit.BaseEntity;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import lombok.Builder;
 import lombok.Getter;
